feat(FOROME-1405): refactoring
