+++ src/pages/filter/common/dashboard/components/body/dashboard-body.module.css
@@ -6,86 +6,4 @@
   justify-content: space-between;
   overflow-y: auto;
   overflow-x: hidden;
-
-  &__tab {
-    display: flex;
-    align-items: center;
-    padding: 0 theme("padding.4");
-    background: theme("colors.blue.darkHover");
-    border-radius: theme("borderRadius.DEFAULT");
-    cursor: grab;
-
-    &__header {
-      @apply typography-m;
-
-      height: 36px;
-      display: flex;
-      flex: 1;
-      align-items: center;
-      justify-content: space-between;
-
-      &__title {
-        font-weight: theme("fontWeight.medium");
-        color: theme("colors.white");
-        cursor: pointer;
-      }
-    }
-  }
-
-  &__sub-tab {
-    @apply typography-s;
-
-    height: auto;
-    min-height: 36px;
-    display: flex;
-    flex-flow: column wrap;
-    align-items: center;
-    padding: 0 theme("padding.4");
-    margin-top: 8px;
-    background: theme("colors.blue.secondary");
-    border-radius: theme("borderRadius.DEFAULT");
-    cursor: grab;
-
-    &__header {
-      height: 36px;
-      width: 100%;
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-
-      &__title {
-        font-weight: theme("fontWeight.medium");
-        color: theme("colors.white");
-        cursor: pointer;
-      }
-    }
-  }
-
-  &__spare-panel {
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    flex-flow: wrap row;
-    padding: 0 theme("padding.4");
-    color: theme("colors.white");
-
-    &__tab {
-      width: 306px;
-      height: 36px;
-      display: flex;
-      align-items: center;
-      margin: theme("margin.1");
-      padding: 0 theme("padding.4");
-      background: theme("colors.blue.darkHover");
-      border-radius: theme("borderRadius.DEFAULT");
-      cursor: pointer;
-
-      &__title {
-        line-height: 3px;
-        font-size: theme("fontSize.16");
-        font-weight: theme("fontWeight.medium");
-        color: theme("colors.white");
-      }
-    }
-  }
 }
