+++ src/pages/filter/common/dashboard/components/body/footer-panel/index.ts
@@ -1,0 +1,1 @@
+export * from './footer-panel'
