+++ src/pages/filter/common/dashboard/components/body/footer-panel/spare-panel.module.css
@@ -1,0 +1,27 @@
+.spare-panel {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  flex-flow: wrap row;
+  padding: 0 theme("padding.4");
+  color: theme("colors.white");
+
+  &__tab {
+    width: 306px;
+    height: 36px;
+    display: flex;
+    align-items: center;
+    margin: theme("margin.1");
+    padding: 0 theme("padding.4");
+    background: theme("colors.blue.darkHover");
+    border-radius: theme("borderRadius.DEFAULT");
+    cursor: pointer;
+
+    &__title {
+      line-height: 3px;
+      font-size: theme("fontSize.16");
+      font-weight: theme("fontWeight.medium");
+      color: theme("colors.white");
+    }
+  }
+}
