+++ src/pages/filter/common/dashboard/components/body/widget-tab/components/widget-sub-tab/index.ts
@@ -1,0 +1,1 @@
+export * from './widget-sub-tab'
