+++ src/pages/filter/common/dashboard/components/body/widget-tab/index.ts
@@ -1,0 +1,1 @@
+export * from './widget-tab'
