+++ src/pages/filter/common/dashboard/components/body/widget-tab/widget-tab.module.css
@@ -1,0 +1,24 @@
+.tab {
+  display: flex;
+  align-items: center;
+  padding: 0 theme("padding.4");
+  background: theme("colors.blue.darkHover");
+  border-radius: theme("borderRadius.DEFAULT");
+  cursor: grab;
+
+  &__header {
+    @apply typography-m;
+
+    height: 36px;
+    display: flex;
+    flex: 1;
+    align-items: center;
+    justify-content: space-between;
+
+    &__title {
+      font-weight: theme("fontWeight.medium");
+      color: theme("colors.white");
+      cursor: pointer;
+    }
+  }
+}
